Reverting the output file feature
The feature has been proven to be a nightmere to be used in real-case
scenarios *and* a bad reimplementation of 'tee', which allows very
easily the same behavior we were looking for. Some changes have
survived though: we're still checking whether the output is to a tty
or not, as non-tty's benefit from newlines more than from '\r'.