Makefile: fix Makefile POSIX-compliance
CPPFLAGS is not defined by POSIX to be used implicitly in .c.o rules,
only CFLAGS is. This commit makes the .c.o rule explicit and uses
CPPFLAGS as such, instead of conflating the contents of CPPFLAGS into
CFLAGS.

This fixes the Makefile for FreeBSD (and presumably, NetBSD as well).